plat/marvell: a80x0: configure Direct Mode window for SPI flash
In order to be able to support UEFI variables, direct mode access
to the SPI flash is required. The variables will be placed at
4MB - 256KB offset. For this purpose at offset 0xf9000000 configure
new window pointing to the CP1 SPI1 CS0 device. Because the platform
supports maximum amount of 6 IO windows, existing window for CP1 PCIE0
was extended to cover the SPI area as well.

Signed-off-by: Marcin Wojtas <mw@semihalf.com>